chore: extract common logic out of server main
